Change the vision detection part to plugin based process, which is more convenient to change among different algorithm of vision filters
